0.4.0
Added support for sounds added in the most recent Thirty Dollar Website update.
Fixed an issue where the icon for the "Alert" sound did not load.
Changed the look of the Sound Select menu to more closely match the Thirty Dollar Website.
Changed the layout of the sounds in the Sound Select menu.
Updated some default sound mappings.